responsive updates beyond Ui design elements
Responsive edits. Please make sure to check out previous versions. I have been constantly updating it to try and get the responsiveness in check while working on current homework. I have had a few tutoring sessions outside of the boot camp and I have another scheduled for this friday with a bootcamp tutor. The grid still needs to be adjusted but the hero image and the nav are now responsive