linked and absurd
linked to case study.

Grids are the STUPIDEST thing ever. When I added the link to the section (because it was the only way to make the link work) it shrunk my image. When I had solved my space issue, the grid undid my work an hour later untouched. I HAVE NO IDEA WHY OR HOW! Why do we need to perfect grids when no one uses them and no professional I currently know can seem to help me fix this including TAs and Tutors. it is absurd that we should be graded on something that the professionals in the field cannot figure out.